app/requirements.txt
autopep8 == 1.5.5
Django == 3.1.7
djangogrpcframework == 0.2
djangorestframework == 3.12.2
grpcio == 1.35.0
grpcio-tools == 1.35.0
kafka-python == 2.0.2
mysqlclient == 2.0.3
pylint == 2.6.2
pytest == 6.2.2
pytest-django == 4.1.0
pytest-grpc == 0.8.0
pytest-mock == 3.5.1
redis == 3.5.3
syrupy == 1.1.0